move frame attributes into a new module
